% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc.R
\docType{package}
\name{bdc}
\alias{bdc}
\title{bdc: An Adapter Library for HPDS Services on the BioData Catalyst Network.}
\description{
This package is used to interact with HPDS Services and data hosted
on a PIC-SURE network.  This library requires the 'picsure' package
to handle all connectivity and authentication with the PIC-SURE
Network itself.
}
