% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc.R
\name{find.in.dictionary}
\alias{find.in.dictionary}
\title{Search the data dictionary of a PIC-SURE resource.}
\usage{
find.in.dictionary(resource, term = "", limit = 0, offset = 0, verbose = FALSE)
}
\arguments{
\item{resource}{A PIC-SURE resource object.}

\item{term}{A string to search for in the resource's data dictionary. By default, if no term is provided, the whole dictionary will be returned.}

\item{limit}{A limit of return results.}

\item{offset}{An offset to start retrning values from.}

\item{verbose}{Flag to display additional runtime information.}
}
\value{
An object representing the search results.
}
\description{
Search the data dictionary of a PIC-SURE resource.
}
\examples{

# myconn <- picsure::connect(url="http://your.server/PIC-SURE/", token="your-security-token")
# myres <- hpds::get.resource(connection=myconn, resourceUUID="YOUR-UUID-0000")
# asthma.terms <- hpds::find.in.dictionary(resource=myres, term="asthma")

}
