% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc.R
\name{get.genotypeAnnotations}
\alias{get.genotypeAnnotations}
\title{Get a list of genomic annotations as a data frame.}
\usage{
get.genotypeAnnotations(resource, verbose = FALSE)
}
\arguments{
\item{resource}{A PIC-SURE resource object.}

\item{verbose}{Flag to display additional runtime information.}
}
\value{
Results in a dataframe object.
}
\description{
Get a list of genomic annotations as a data frame.
}
\examples{

# myconn <- picsure::connect(url="http://your.server/PIC-SURE/", token="your-security-token")
# myres <- hpds::get.resource(connection=myconn, resourceUUID="YOUR-UUID-0000")
# get.genotypeAnnotations(myres)

}
