% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc.R
\name{get.paths}
\alias{get.paths}
\title{Get a list of HPDS paths.}
\usage{
get.paths(dictionary.results, verbose = FALSE)
}
\arguments{
\item{dictionary.results}{A data dictionary search results object.}

\item{verbose}{Flag to display additional runtime information.}
}
\value{
A list of HPDS paths.
}
\description{
Get a list of HPDS paths.
}
\examples{

# myconn <- picsure::connect(url="http://your.server/PIC-SURE/", token="your-security-token")
# myres <- hpds::get.resource(connection=myconn, resourceUUID="YOUR-UUID-0000")
# asthma.terms <- hpds::find.in.dictionary(resource=myres, term="asthma")
# get.paths(asthma.terms)

}
