% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc.R
\name{get.varInfo}
\alias{get.varInfo}
\title{Get variable info for a given path.}
\usage{
get.varInfo(dictionary.results, path, verbose = FALSE)
}
\arguments{
\item{dictionary.results}{A data dictionary search results object.}

\item{path}{A character object containing an HPDS path.}

\item{verbose}{Flag to display additional runtime information.}
}
\value{
A data frame object containing variable info for an HPDS path.
}
\description{
Get variable info for a given path.
}
\examples{

# myconn <- picsure::connect(url="http://your.server/PIC-SURE/", token="your-security-token")
# myres <- hpds::get.resource(connection=myconn, resourceUUID="YOUR-UUID-0000")
# asthma.terms <- hpds::find.in.dictionary(resource=myres, term="asthma")
# get.varInfo(asthma.terms, '\\phs001211\\...')

}
