% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-classes.R
\docType{class}
\name{PicSureHpdsDictionaryBdc}
\alias{PicSureHpdsDictionaryBdc}
\title{R6 class that runs searches against a HPDS resource's data dictionary - DO NOT CREATE THIS OBJECT DIRECTLY!
For full documentation of each method go to https://github.com/hms-dbmi/pic-sure-r-adapter-hpds}
\format{
\code{\link{PicSureHpdsDictionary}} object.
}
\value{
Object of \code{\link{R6Class}} used to access a HPDS-hosted resource's data dictionary.
}
\description{
R6 class that runs searches against a HPDS resource's data dictionary - DO NOT CREATE THIS OBJECT DIRECTLY!
For full documentation of each method go to https://github.com/hms-dbmi/pic-sure-r-adapter-hpds

R6 class that runs searches against a HPDS resource's data dictionary - DO NOT CREATE THIS OBJECT DIRECTLY!
For full documentation of each method go to https://github.com/hms-dbmi/pic-sure-r-adapter-hpds
}
\keyword{data}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PicSureHpdsDictionaryBdc-new}{\code{PicSureHpdsDictionaryBdc$new()}}
\item \href{#method-PicSureHpdsDictionaryBdc-getKeyInfo}{\code{PicSureHpdsDictionaryBdc$getKeyInfo()}}
\item \href{#method-PicSureHpdsDictionaryBdc-genotypeAnnotations}{\code{PicSureHpdsDictionaryBdc$genotypeAnnotations()}}
\item \href{#method-PicSureHpdsDictionaryBdc-find}{\code{PicSureHpdsDictionaryBdc$find()}}
\item \href{#method-PicSureHpdsDictionaryBdc-clone}{\code{PicSureHpdsDictionaryBdc$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PicSureHpdsDictionaryBdc-new"></a>}}
\if{latex}{\out{\hypertarget{method-PicSureHpdsDictionaryBdc-new}{}}}
\subsection{Method \code{new()}}{
Populate a hash/dictionary of phenotype and genotype concepts.
This method is used to create new object of this class. DO NOT CREATE THIS OBJECT DIRECTLY!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsDictionaryBdc$new(resourceConnection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resourceConnection}}{A `PicSureHpdsResourceConnectionBdc` object.
Helper method to return key/path information from phenotype concept paths.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PicSureHpdsDictionaryBdc-getKeyInfo"></a>}}
\if{latex}{\out{\hypertarget{method-PicSureHpdsDictionaryBdc-getKeyInfo}{}}}
\subsection{Method \code{getKeyInfo()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsDictionaryBdc$getKeyInfo(key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The concept path key to look up.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Concept path information or FALSE.
Returns genotype annotations from info concept paths.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PicSureHpdsDictionaryBdc-genotypeAnnotations"></a>}}
\if{latex}{\out{\hypertarget{method-PicSureHpdsDictionaryBdc-genotypeAnnotations}{}}}
\subsection{Method \code{genotypeAnnotations()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsDictionaryBdc$genotypeAnnotations()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
a `Data.Frame` object containing a table of genotype annotations.
This method returns a `PicSureHpdsDictionaryBdcResult` object containing the results of the search on the HPDS resource's data dictionary.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PicSureHpdsDictionaryBdc-find"></a>}}
\if{latex}{\out{\hypertarget{method-PicSureHpdsDictionaryBdc-find}{}}}
\subsection{Method \code{find()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsDictionaryBdc$find(term, limit, offset, showAll = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{term}}{Term to find in the dictionary.}

\item{\code{limit}}{Number of items to return.}

\item{\code{offset}}{The offset record to start returning from.}

\item{\code{showAll}}{Show all studies, not just ones within connection query scopes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `PicSureHpdsDictionaryBdcResult` object containing results from the dictionary.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PicSureHpdsDictionaryBdc-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PicSureHpdsDictionaryBdc-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsDictionaryBdc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
