% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-classes.R
\docType{class}
\name{PicSureHpdsResourceConnectionBdc}
\alias{PicSureHpdsResourceConnectionBdc}
\title{R6 class that allows access to the data dictionary and query services of a selected HPDS-hosted resources on a PIC-SURE network.}
\format{
\code{\link{PicSureHpdsResourceConnection}} object.
}
\value{
Object of \code{\link{R6Class}} used to access the dictionary and query services via the objects it returns.
}
\description{
R6 class that allows access to the data dictionary and query services of a selected HPDS-hosted resources on a PIC-SURE network.

R6 class that allows access to the data dictionary and query services of a selected HPDS-hosted resources on a PIC-SURE network.
}
\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/hms-dbmi/pic-sure-r-adapter-hpds}
  \item{\code{new(connection, resource_uuid)}}{This method is used to create new object of this class which uses the passed PicSureConnection object for communication with the PIC-SURE network along with a UUID to identify a HPDS-hosted resource.}

  \item{\code{dictionary()}}{This method returns a \code{PicSureHpdsDictionary} object which is used to run lookups against a resources data dictionary.}
  \item{\code{query()}}{This method returns a new \code{PicSureHpdsQuery} object configured to run all commands against the previously specified HPDS-hosted resource.}}
}

\keyword{data}
\section{Super class}{
\code{\link[hpds:PicSureHpdsResourceConnection]{hpds::PicSureHpdsResourceConnection}} -> \code{PicSureHpdsResourceConnectionBdc}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PicSureHpdsResourceConnectionBdc$new()}}
\item \href{#method-query}{\code{PicSureHpdsResourceConnectionBdc$query()}}
\item \href{#method-clone}{\code{PicSureHpdsResourceConnectionBdc$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="hpds" data-topic="PicSureHpdsResourceConnection" data-id="dictionary">}\href{../../hpds/html/PicSureHpdsResourceConnection.html#method-dictionary}{\code{hpds::PicSureHpdsResourceConnection$dictionary()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hpds" data-topic="PicSureHpdsResourceConnection" data-id="getQueryByUUID">}\href{../../hpds/html/PicSureHpdsResourceConnection.html#method-getQueryByUUID}{\code{hpds::PicSureHpdsResourceConnection$getQueryByUUID()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hpds" data-topic="PicSureHpdsResourceConnection" data-id="retrieveQueryResults">}\href{../../hpds/html/PicSureHpdsResourceConnection.html#method-retrieveQueryResults}{\code{hpds::PicSureHpdsResourceConnection$retrieveQueryResults()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hpds" data-topic="PicSureHpdsResourceConnection" data-id="version">}\href{../../hpds/html/PicSureHpdsResourceConnection.html#method-version}{\code{hpds::PicSureHpdsResourceConnection$version()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsResourceConnectionBdc$new(connection, resource_uuid)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-query"></a>}}
\if{latex}{\out{\hypertarget{method-query}{}}}
\subsection{Method \code{query()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsResourceConnectionBdc$query(loadQuery = NA)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsResourceConnectionBdc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
