% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc.R
\name{use.dictionary}
\alias{use.dictionary}
\title{Get a new reference to the dictionary HPDS-based PIC-SURE resource.}
\usage{
use.dictionary(connection, verbose = FALSE)
}
\arguments{
\item{connection}{A PIC-SURE connection object.}

\item{verbose}{Flag to display additional runtime information.}
}
\value{
An object which provides access to the dictionary resource.
}
\description{
Get a new reference to the dictionary HPDS-based PIC-SURE resource.
}
\examples{

# myconn <- picsure::connect(url="http://your.server/PIC-SURE/", token="your-security-token")
# myres <- hpds::use.dictionary(connection=myconn)

}
