% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdc.R
\name{use.openPicSure}
\alias{use.openPicSure}
\title{Get a new reference to the open PIC-SURE HPDS-based resource.}
\usage{
use.openPicSure(connection, verbose = FALSE)
}
\arguments{
\item{connection}{A PIC-SURE connection object.}

\item{verbose}{Flag to display additional runtime information.}
}
\value{
An object which provides access to the open PIC-SURE resource.
}
\description{
Get a new reference to the open PIC-SURE HPDS-based resource.
}
\examples{

# myconn <- picsure::connect(url="http://your.server/PIC-SURE/", token="your-security-token")
# myres <- hpds::use.openPicSure(connection=myconn)

}
