% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-classes.R
\docType{class}
\name{PicSureHpdsDictionaryBdcResult}
\alias{PicSureHpdsDictionaryBdcResult}
\title{PicSureHpdsDictionaryBdcResult}
\value{
Object of \code{\link{R6Class}} used to access a HPDS-hosted resource's data dictionary.
}
\description{
R6 class contain the results of a search against a HPDS resource's data dictionary - DO NOT CREATE THIS OBJECT DIRECTLY!

For full documentation of each method go to https://github.com/hms-dbmi/pic-sure-r-adapter-hpds
}
\keyword{data}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PicSureHpdsDictionaryBdcResult-new}{\code{PicSureHpdsDictionaryBdcResult$new()}}
\item \href{#method-PicSureHpdsDictionaryBdcResult-count}{\code{PicSureHpdsDictionaryBdcResult$count()}}
\item \href{#method-PicSureHpdsDictionaryBdcResult-entries}{\code{PicSureHpdsDictionaryBdcResult$entries()}}
\item \href{#method-PicSureHpdsDictionaryBdcResult-varInfo}{\code{PicSureHpdsDictionaryBdcResult$varInfo()}}
\item \href{#method-PicSureHpdsDictionaryBdcResult-paths}{\code{PicSureHpdsDictionaryBdcResult$paths()}}
\item \href{#method-PicSureHpdsDictionaryBdcResult-dataframe}{\code{PicSureHpdsDictionaryBdcResult$dataframe()}}
\item \href{#method-PicSureHpdsDictionaryBdcResult-clone}{\code{PicSureHpdsDictionaryBdcResult$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PicSureHpdsDictionaryBdcResult-new"></a>}}
\if{latex}{\out{\hypertarget{method-PicSureHpdsDictionaryBdcResult-new}{}}}
\subsection{Method \code{new()}}{
Project and filter fields in query scope from results, mapping them with new field names.


This method is used to create PicSureHpdsDictionaryBdcResult object. DO NOT CREATE THIS OBJECT DIRECTLY!
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsDictionaryBdcResult$new(results, scopes, showAll = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{A list of results to process and return.}

\item{\code{scopes}}{A list of query scopes for filter by.}

\item{\code{showAll}}{Show all studies, not just ones within connection query scopes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list that contains paths and filtered results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PicSureHpdsDictionaryBdcResult-count"></a>}}
\if{latex}{\out{\hypertarget{method-PicSureHpdsDictionaryBdcResult-count}{}}}
\subsection{Method \code{count()}}{
This method returns a integer of how many terms were returned by the data dictionary search.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsDictionaryBdcResult$count()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Return the number of returned results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PicSureHpdsDictionaryBdcResult-entries"></a>}}
\if{latex}{\out{\hypertarget{method-PicSureHpdsDictionaryBdcResult-entries}{}}}
\subsection{Method \code{entries()}}{
This method returns information about the terms discovered by the data dictionary search in a data frame format.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsDictionaryBdcResult$entries()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Return all results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PicSureHpdsDictionaryBdcResult-varInfo"></a>}}
\if{latex}{\out{\hypertarget{method-PicSureHpdsDictionaryBdcResult-varInfo}{}}}
\subsection{Method \code{varInfo()}}{
Display all information available on a particular HPDS_PATH corresponding to a search result.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsDictionaryBdcResult$varInfo(path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{A path to look for in the results.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `Data.Frame` containing the fields for the given path.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PicSureHpdsDictionaryBdcResult-paths"></a>}}
\if{latex}{\out{\hypertarget{method-PicSureHpdsDictionaryBdcResult-paths}{}}}
\subsection{Method \code{paths()}}{
Extract the paths of all results into a list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsDictionaryBdcResult$paths()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list of paths returned from the results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PicSureHpdsDictionaryBdcResult-dataframe"></a>}}
\if{latex}{\out{\hypertarget{method-PicSureHpdsDictionaryBdcResult-dataframe}{}}}
\subsection{Method \code{dataframe()}}{
Flatten the results into a dataframe.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsDictionaryBdcResult$dataframe()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A `Data.Frame` of all results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PicSureHpdsDictionaryBdcResult-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PicSureHpdsDictionaryBdcResult-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsDictionaryBdcResult$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
