% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R6-classes.R
\docType{class}
\name{PicSureHpdsQueryBdc}
\alias{PicSureHpdsQueryBdc}
\title{R6 class used to build a multi-use query to search against a HPDS resource's data - DO NOT CREATE THIS OBJECT DIRECTLY!
For full documentation of each method go to https://github.com/hms-dbmi/pic-sure-r-adapter-hpds}
\format{
\code{\link{PicSureHpdsQuery}} object.
This method is used to create new `PicSureHpdsQueryBdc` object. DO NOT CREATE THIS OBJECT DIRECTLY!
}
\value{
Object of \code{\link{R6Class}} used to access a HPDS-hosted resource's data dictionary.
}
\description{
R6 class used to build a multi-use query to search against a HPDS resource's data - DO NOT CREATE THIS OBJECT DIRECTLY!
For full documentation of each method go to https://github.com/hms-dbmi/pic-sure-r-adapter-hpds

R6 class used to build a multi-use query to search against a HPDS resource's data - DO NOT CREATE THIS OBJECT DIRECTLY!
For full documentation of each method go to https://github.com/hms-dbmi/pic-sure-r-adapter-hpds
}
\keyword{data}
\section{Super class}{
\code{\link[hpds:PicSureHpdsQuery]{hpds::PicSureHpdsQuery}} -> \code{PicSureHpdsQueryBdc}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PicSureHpdsQueryBdc-new}{\code{PicSureHpdsQueryBdc$new()}}
\item \href{#method-PicSureHpdsQueryBdc-buildQuery}{\code{PicSureHpdsQueryBdc$buildQuery()}}
\item \href{#method-PicSureHpdsQueryBdc-clone}{\code{PicSureHpdsQueryBdc$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="hpds" data-topic="PicSureHpdsQuery" data-id="anyof"><a href='../../hpds/html/PicSureHpdsQuery.html#method-PicSureHpdsQuery-anyof'><code>hpds::PicSureHpdsQuery$anyof()</code></a></li>
<li><span class="pkg-link" data-pkg="hpds" data-topic="PicSureHpdsQuery" data-id="crosscounts"><a href='../../hpds/html/PicSureHpdsQuery.html#method-PicSureHpdsQuery-crosscounts'><code>hpds::PicSureHpdsQuery$crosscounts()</code></a></li>
<li><span class="pkg-link" data-pkg="hpds" data-topic="PicSureHpdsQuery" data-id="filter"><a href='../../hpds/html/PicSureHpdsQuery.html#method-PicSureHpdsQuery-filter'><code>hpds::PicSureHpdsQuery$filter()</code></a></li>
<li><span class="pkg-link" data-pkg="hpds" data-topic="PicSureHpdsQuery" data-id="getCount"><a href='../../hpds/html/PicSureHpdsQuery.html#method-PicSureHpdsQuery-getCount'><code>hpds::PicSureHpdsQuery$getCount()</code></a></li>
<li><span class="pkg-link" data-pkg="hpds" data-topic="PicSureHpdsQuery" data-id="getResults"><a href='../../hpds/html/PicSureHpdsQuery.html#method-PicSureHpdsQuery-getResults'><code>hpds::PicSureHpdsQuery$getResults()</code></a></li>
<li><span class="pkg-link" data-pkg="hpds" data-topic="PicSureHpdsQuery" data-id="getResultsDataFrame"><a href='../../hpds/html/PicSureHpdsQuery.html#method-PicSureHpdsQuery-getResultsDataFrame'><code>hpds::PicSureHpdsQuery$getResultsDataFrame()</code></a></li>
<li><span class="pkg-link" data-pkg="hpds" data-topic="PicSureHpdsQuery" data-id="getRunDetails"><a href='../../hpds/html/PicSureHpdsQuery.html#method-PicSureHpdsQuery-getRunDetails'><code>hpds::PicSureHpdsQuery$getRunDetails()</code></a></li>
<li><span class="pkg-link" data-pkg="hpds" data-topic="PicSureHpdsQuery" data-id="getVariantsApproximateCount"><a href='../../hpds/html/PicSureHpdsQuery.html#method-PicSureHpdsQuery-getVariantsApproximateCount'><code>hpds::PicSureHpdsQuery$getVariantsApproximateCount()</code></a></li>
<li><span class="pkg-link" data-pkg="hpds" data-topic="PicSureHpdsQuery" data-id="getVariantsDataFrame"><a href='../../hpds/html/PicSureHpdsQuery.html#method-PicSureHpdsQuery-getVariantsDataFrame'><code>hpds::PicSureHpdsQuery$getVariantsDataFrame()</code></a></li>
<li><span class="pkg-link" data-pkg="hpds" data-topic="PicSureHpdsQuery" data-id="load"><a href='../../hpds/html/PicSureHpdsQuery.html#method-PicSureHpdsQuery-load'><code>hpds::PicSureHpdsQuery$load()</code></a></li>
<li><span class="pkg-link" data-pkg="hpds" data-topic="PicSureHpdsQuery" data-id="loadInternal"><a href='../../hpds/html/PicSureHpdsQuery.html#method-PicSureHpdsQuery-loadInternal'><code>hpds::PicSureHpdsQuery$loadInternal()</code></a></li>
<li><span class="pkg-link" data-pkg="hpds" data-topic="PicSureHpdsQuery" data-id="require"><a href='../../hpds/html/PicSureHpdsQuery.html#method-PicSureHpdsQuery-require'><code>hpds::PicSureHpdsQuery$require()</code></a></li>
<li><span class="pkg-link" data-pkg="hpds" data-topic="PicSureHpdsQuery" data-id="save"><a href='../../hpds/html/PicSureHpdsQuery.html#method-PicSureHpdsQuery-save'><code>hpds::PicSureHpdsQuery$save()</code></a></li>
<li><span class="pkg-link" data-pkg="hpds" data-topic="PicSureHpdsQuery" data-id="select"><a href='../../hpds/html/PicSureHpdsQuery.html#method-PicSureHpdsQuery-select'><code>hpds::PicSureHpdsQuery$select()</code></a></li>
<li><span class="pkg-link" data-pkg="hpds" data-topic="PicSureHpdsQuery" data-id="show"><a href='../../hpds/html/PicSureHpdsQuery.html#method-PicSureHpdsQuery-show'><code>hpds::PicSureHpdsQuery$show()</code></a></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PicSureHpdsQueryBdc-new"></a>}}
\if{latex}{\out{\hypertarget{method-PicSureHpdsQueryBdc-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsQueryBdc$new(connection)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connection}}{A `PicSureHpdsResourceConnectionBdc` object.
This method is used to create new `PicSureHpdsQueryBdc` object. DO NOT CREATE THIS OBJECT DIRECTLY!}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PicSureHpdsQueryBdc-buildQuery"></a>}}
\if{latex}{\out{\hypertarget{method-PicSureHpdsQueryBdc-buildQuery}{}}}
\subsection{Method \code{buildQuery()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsQueryBdc$buildQuery(resultType = "COUNT")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{resultType}}{The type of result that should be returned from the api.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object to use for the query.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PicSureHpdsQueryBdc-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PicSureHpdsQueryBdc-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PicSureHpdsQueryBdc$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
